# Improve performance of `Reline::Unicode.get_mbchar_width`