@mblumtritt 2024-08-16
Sorry for another question: Am I right that country flags (like '🇵🇹') or not supported?