@tompng 2024-08-16
> Am I right that country flags (like '🇵🇹') are not supported?

In my opinion, Reline supports calculating flag width correctly.
Flags are defined as N(Neutral == narrow or half), width=1 character in Unicode EastAsianWidth.txt.
```
# EastAsianWidth.txt
1F1E6..1F1FF   ; N  # So    [26] REGIONAL INDICATOR SYMBOL LETTER A..REGIONAL INDICATOR SYMBOL LETTER Z
```

But flag rendering fails in many terminal emulators with various reasons.

### Mac Terminal.app
Renders with half-width visually but calculates column as full-width internally. Can't support a bug.

### iTerm2
Flag width can be changed by experimental configuration `Fag emoji render full-width: yes/no` and the default is full-width.
Width of some half-width emojis(example: ⛴️) are also configurable.
There is no plan to detect this kind of configuration value from Reline.

### Alacritty, VSCode Terminal
Does not support rendering flags. "🇵🇹" is rendered as "P"+"T"