@elfham 2024-08-17
IMHO, emojis of flags should be treated as double-width characters if they are combined in the same way as half-width katakana characters.

```
% irb
irb(main):001> Reline::Unicode.get_mbchar_width("ｶ")
=> 1
irb(main):002> Reline::Unicode.get_mbchar_width("ｶﾞ")
=> 2
irb(main):003> 
```

However, this does not seem to be well documented in UAX#11 and is handled differently by different implementations.

I think there is still room for further discussion.

So, I think it would be better to discuss the flag width as a separate Issue from this PR.