@mblumtritt 2024-08-17
Ok. thanks for your answer!

Let's have this PR merged asap - which will help a lot to make things better. And maybe we handle the issue with some special chars which are different handled in terminals later…

Thanks!